requirements.txt
Django==1.6.2
Markdown==2.4
Pygments==1.6
South==0.8.4
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-toolbelt==0.0.1
djangorestframework==2.3.12
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
